refactor ui components to a ui module
